requirements.txt
# 3Dモデル-画像マッチング 必要ライブラリ
# インストール: pip install -r requirements.txt

# Core dependencies
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
tqdm>=4.60.0

# 3D processing
trimesh>=3.15.0
pyrender>=0.1.45
networkx>=2.6.0

# Database
# SQLite3 is included in Python standard library

# Optional: For interactive viewer
# flask>=2.0.0
# flask-cors>=3.0.10
